requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.1
cryptography>=44.0.0
python-multipart>=0.0.12
httpx>=0.28.0
redis>=5.2.0
arq>=0.26.1
google-api-python-client>=2.154.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
anthropic>=0.40.0
email-validator>=2.2.0
aiofiles>=24.1.0
jinja2>=3.1.4
